fix proto conflict
