+++ federatedrec/factorization_machine/hetero_factorization_machine/hetero_fm_base.py
@@ -20,7 +20,7 @@ import numpy as np
 
 from federatedml.util import consts
 from federatedml.secureprotol import PaillierEncrypt
-from federatedrec.protobuf.generated import fm_model_meta_pb2
+from federatedml.protobuf.generated import fm_model_meta_pb2
 from federatedrec.param.factorization_machine_param import HeteroFactorizationParam
 from federatedrec.factorization_machine.base_factorization_machine import BaseFactorizationMachine
 from federatedrec.transfer_variable.transfer_class.hetero_fm_transfer_variable import HeteroFMTransferVariable
