--- federatedrec/protobuf/proto/fm-model-meta.proto
@@ -1,36 +1,0 @@
-/*
- * Copyright 2019 The FATE Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-syntax = "proto3";
-
-package com.webank.ai.fate.core.mlmodel.buffer.fm;
-option java_outer_classname = "FMModelMetaProto";
-
-
-message FMModelMeta {
-    string penalty = 1;
-    double tol = 2;
-    double alpha = 3;
-    string optimizer = 4;
-    double party_weight = 5;
-    int64 batch_size = 6;
-    double learning_rate = 7;
-    int64 max_iter = 8;
-    string early_stop = 9;
-    int64 re_encrypt_batches = 10;
-    bool fit_intercept=11;
-    bool need_one_vs_rest = 12;
-}
